requirements.txt
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask==0.12.2
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
PyYAML==3.12
SQLAlchemy==1.2.3
Werkzeug==0.14.1
alembic==0.9.8
aniso8601==2.0.0
apispec==0.31.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
flask-marshmallow==0.8.0
flask-restplus-patched==0.1.3
flask-restplus==0.10.1
idna==2.6
itsdangerous==0.24
jsonschema==2.6.0
marshmallow==2.15.0
python-dateutil==2.6.1
python-decouple==3.1
python-editor==1.0.3
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
webargs==2.0.0
